requirements.txt
streamlit==1.47.1
pandas==2.3.1
numpy==2.3.2
altair==5.5.0
plotly==6.2.0
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.7.1
statsmodels==0.14.5

# BigQuery + auth
google-cloud-bigquery==3.25.0
db-dtypes==1.3.0
pyarrow==21.0.0
google-auth==2.35.0